requirement.txt
Flask
werkzeug
gunicorn
mysql-connector-python
bcrypt
scikit-learn
flask-cors

